+++ reference_data/models.py
@@ -239,3 +239,13 @@ class MGI(models.Model):
     class Meta:
         unique_together = ('gene', 'marker_id')
         json_fields = ['marker_id']
+
+
+class GenCC(models.Model):
+    gene = models.ForeignKey(GeneInfo, on_delete=models.CASCADE)
+
+    hgnc_id = models.CharField(max_length=10)
+    classifications = models.JSONField()
+
+    class Meta:
+        json_fields = ['classifications', 'hgnc_id']