+++ ui/pages/Project/components/Matchmaker.jsx
@@ -363,6 +363,24 @@ const DISPLAY_FIELDS = [
   },
 ]
 
+const UpdateMmeSubmissionButton = ({ individual, action, ...props }) => (
+  <UpdateButton
+    modalSize="large"
+    modalTitle={`${action} Submission for ${individual.displayName}`}
+    modalId={`${individual.individualGuid}_-_${action}MmeSubmission`}
+    formMetaId={individual.individualGuid}
+    confirmDialog={`Are you sure you want to ${action === 'Create' ? 'submit this individual' : 'update this submission'}?`}
+    formFields={SUBMISSION_EDIT_FIELDS}
+    showErrorPanel
+    {...props}
+  />
+)
+
+UpdateMmeSubmissionButton.propTypes = {
+  individual: PropTypes.object.isRequired,
+  action: PropTypes.string,
+}
+
 const BaseMatchmakerIndividual = React.memo((
   { loading, load, searchMme, individual, onSubmit, defaultMmeSubmission, mmeResults, mmeSubmission },
 ) => (
@@ -399,18 +417,13 @@ const BaseMatchmakerIndividual = React.memo((
           icon={<Icon name="warning sign" color="orange" />}
           subheader={
             <div className="sub header">
-              <UpdateButton
+              <UpdateMmeSubmissionButton
+                action="Create"
+                individual={individual}
                 initialValues={defaultMmeSubmission}
                 buttonText="Submit to Matchmaker"
                 editIconName=" "
-                modalSize="large"
-                modalTitle={`Create Submission for ${individual.displayName}`}
-                modalId={`${individual.individualGuid}_-_createMmeSubmission`}
-                formMetaId={individual.individualGuid}
-                confirmDialog="Are you sure you want to submit this individual?"
-                formFields={SUBMISSION_EDIT_FIELDS}
                 onSubmit={onSubmit}
-                showErrorPanel
               />
             </div>
           }
@@ -428,17 +441,13 @@ const BaseMatchmakerIndividual = React.memo((
             content="Search for New Matches"
           />
           | &nbsp; &nbsp;
-          <UpdateButton
+          <UpdateMmeSubmissionButton
+            action="Update"
+            individual={individual}
             disabled={loading}
             buttonText="Update Submission"
-            modalSize="large"
-            modalTitle={`Update Submission for ${individual.displayName}`}
-            modalId={`${individual.individualGuid}_-_updateMmeSubmission`}
-            confirmDialog="Are you sure you want to update this submission?"
             initialValues={mmeSubmission}
-            formFields={SUBMISSION_EDIT_FIELDS}
             onSubmit={onSubmit}
-            showErrorPanel
           />
           | &nbsp; &nbsp;
           <DeleteButton