+++ ui/redux/selectors.js
@@ -83,6 +83,16 @@ export const getNotesByFamilyType = createSelector(
   }, {}),
 )
 
+export const getProjectAnalysisGroupOptions = createSelector(
+  getAnalysisGroupsGroupedByProjectGuid,
+  analysisGroupsByProject => Object.entries(analysisGroupsByProject).reduce(
+    (acc, [projectGuid, analysisGroupsByGuid]) => ({
+      ...acc,
+      [projectGuid]: Object.values(analysisGroupsByGuid).sort((a, b) => a.name.localeCompare(b.name)),
+    }), {},
+  ),
+)
+
 export const getAnalysisGroupsByFamily = createSelector(
   getAnalysisGroupsByGuid,
   analysisGroupsByGuid => Object.values(analysisGroupsByGuid).reduce(